22 Wantage Close
"What are you going to do, then?" I asked.

"To smoke," he answered. "It is quite a three pipe problem, and I beg that you won't speak to me for
fifty minutes." He curled himself up in his chair, with his thin knees drawn up to his hawk-like
nose, and there he sat with his eyes closed and his black clay pipe thrusting out like the bill of
some strange bird. I had come to the conclusion that he had dropped asleep, and indeed was nodding
myself, when he suddenly sprang out of his chair with the gesture of a man who has made up his mind
and put his pipe down upon the mantelpiece.

"Sarasate plays at the St. James's Hall this afternoon," he remarked. "What do you think, Watson?
Could your patients spare you for a few hours?"

"I have nothing to do to-day. My practice is never very absorbing."

"Then put on your hat and come. I am going through the City first, and we can have some lunch on the
way. I observe that there is a good deal of German music on the programme, which is rather more to
my taste than Italian or French. It is introspective, and I want to introspect. Come along!"

We travelled by the Underground as far as Aldersgate; and a short walk took us to Saxe-Coburg
Square, the scene of the singular story which we had listened to in the morning. It was a poky,
little, shabby-genteel place, where four lines of dingy two-storied brick houses looked out into a
small railed-in enclosure, where a lawn of weedy grass and a few clumps of faded laurel-bushes made
a hard fight against a smoke-laden and uncongenial atmosphere. Three gilt balls and a brown board
with "Jabez Wilson" in white letters, upon a corner house, announced the place where our red-headed
client carried on his business. Sherlock Holmes stopped in front of it with his head on one side and
looked it all over, with his eyes shining brightly between puckered lids. Then he walked slowly up
the street, and then down again to the corner, still looking keenly at the houses. Finally he
returned to the pawnbroker's, and, having thumped vigorously upon the pavement with his stick two or
three times, he went up to the door and knocked. It was instantly opened by a bright-looking, clean-
shaven young fellow, who asked him to step in.