143 Glan Road
"Well, to come to an end of the matter, Mr. Holmes, and not to abuse your patience, there came a
night when he made one of those drunken sallies from which he never came back. We found him, when we
went to search for him, face downward in a little green-scummed pool, which lay at the foot of the
garden. There was no sign of any violence, and the water was but two feet deep, so that the jury,
having regard to his known eccentricity, brought in a verdict of 'suicide.' But I, who knew how he
winced from the very thought of death, had much ado to persuade myself that he had gone out of his
way to meet it. The matter passed, however, and my father entered into possession of the estate, and
of some £14,000, which lay to his credit at the bank."

"One moment," Holmes interposed, "your statement is, I foresee, one of the most remarkable to which
I have ever listened. Let me have the date of the reception by your uncle of the letter, and the
date of his supposed suicide."

"The letter arrived on March 10, 1883. His death was seven weeks later, upon the night of May 2nd."