132 Harvey Street
"Hum! We will talk about that. And now let us hear a true account of the next act. How came the
stone into the goose, and how came the goose into the open market? Tell us the truth, for there lies
your only hope of safety."

Ryder passed his tongue over his parched lips. "I will tell you it just as it happened, sir," said
he. "When Horner had been arrested, it seemed to me that it would be best for me to get away with
the stone at once, for I did not know at what moment the police might not take it into their heads
to search me and my room. There was no place about the hotel where it would be safe. I went out, as
if on some commission, and I made for my sister's house. She had married a man named Oakshott, and
lived in Brixton Road, where she fattened fowls for the market. All the way there every man I met
seemed to me to be a policeman or a detective; and, for all that it was a cold night, the sweat was
pouring down my face before I came to the Brixton Road. My sister asked me what was the matter, and
why I was so pale; but I told her that I had been upset by the jewel robbery at the hotel. Then I
went into the back yard and smoked a pipe and wondered what it would be best to do.

"I had a friend once called Maudsley, who went to the bad, and has just been serving his time in
Pentonville. One day he had met me, and fell into talk about the ways of thieves, and how they could
get rid of what they stole. I knew that he would be true to me, for I knew one or two things about
him; so I made up my mind to go right on to Kilburn, where he lived, and take him into my
confidence. He would show me how to turn the stone into money. But how to get to him in safety? I
thought of the agonies I had gone through in coming from the hotel. I might at any moment be seized
and searched, and there would be the stone in my waistcoat pocket. I was leaning against the wall at
the time and looking at the geese which were waddling about round my feet, and suddenly an idea came
into my head which showed me how I could beat the best detective that ever lived.

"My sister had told me some weeks before that I might have the pick of her geese for a Christmas
present, and I knew that she was always as good as her word. I would take my goose now, and in it I
would carry my stone to Kilburn. There was a little shed in the yard, and behind this I drove one of
the birds--a fine big one, white, with a barred tail. I caught it, and prying its bill open, I
thrust the stone down its throat as far as my finger could reach. The bird gave a gulp, and I felt
the stone pass along its gullet and down into its crop. But the creature flapped and struggled, and
out came my sister to know what was the matter. As I turned to speak to her the brute broke loose
and fluttered off among the others.

"'Whatever were you doing with that bird, Jem?' says she.