106 Down Close
"'The firm does so when the security is good.' I answered.

"'It is absolutely essential to me,' said he, 'that I should have £50,000 at once. I could, of
course, borrow so trifling a sum ten times over from my friends, but I much prefer to make it a
matter of business and to carry out that business myself. In my position you can readily understand
that it is unwise to place one's self under obligations.'