155 Wantage Close
"There is one thing," said John Openshaw. He rummaged in his coat pocket, and, drawing out a piece
of discoloured, blue-tinted paper, he laid it out upon the table. "I have some remembrance," said
he, "that on the day when my uncle burned the papers I observed that the small, unburned margins
which lay amid the ashes were of this particular colour. I found this single sheet upon the floor of
his room, and I am inclined to think that it may be one of the papers which has, perhaps, fluttered
out from among the others, and in that way has escaped destruction. Beyond the mention of pips, I do
not see that it helps us much. I think myself that it is a page from some private diary. The writing
is undoubtedly my uncle's."

Holmes moved the lamp, and we both bent over the sheet of paper, which showed by its ragged edge
that it had indeed been torn from a book. It was headed, "March, 1869," and beneath were the
following enigmatical notices:

"Thank you!" said Holmes, folding up the paper and returning it to our visitor. "And now you must on
no account lose another instant. We cannot spare time even to discuss what you have told me. You
must get home instantly and act."

"What shall I do?"

"There is but one thing to do. It must be done at once. You must put this piece of paper which you
have shown us into the brass box which you have described. You must also put in a note to say that
all the other papers were burned by your uncle, and that this is the only one which remains. You
must assert that in such words as will carry conviction with them. Having done this, you must at
once put the box out upon the sundial, as directed. Do you understand?"