12 Cliff Place
"I felt angry at having been tricked by so elaborate a story as that which he had told me. 'I was
admiring your fuller's-earth,' said I; 'I think that I should be better able to advise you as to
your machine if I knew what the exact purpose was for which it was used.'

"The instant that I uttered the words I regretted the rashness of my speech. His face set hard, and
a baleful light sprang up in his grey eyes.

"'Very well,' said he, 'you shall know all about the machine.' He took a step backward, slammed the
little door, and turned the key in the lock. I rushed towards it and pulled at the handle, but it
was quite secure, and did not give in the least to my kicks and shoves. 'Hullo!' I yelled. 'Hullo!
Colonel! Let me out!'

"And then suddenly in the silence I heard a sound which sent my heart into my mouth. It was the
clank of the levers and the swish of the leaking cylinder. He had set the engine at work. The lamp
still stood upon the floor where I had placed it when examining the trough. By its light I saw that
the black ceiling was coming down upon me, slowly, jerkily, but, as none knew better than myself,
with a force which must within a minute grind me to a shapeless pulp. I threw myself, screaming,
against the door, and dragged with my nails at the lock. I implored the colonel to let me out, but
the remorseless clanking of the levers drowned my cries. The ceiling was only a foot or two above my
head, and with my hand upraised I could feel its hard, rough surface. Then it flashed through my
mind that the pain of my death would depend very much upon the position in which I met it. If I lay
on my face the weight would come upon my spine, and I shuddered to think of that dreadful snap.
Easier the other way, perhaps; and yet, had I the nerve to lie and look up at that deadly black
shadow wavering down upon me? Already I was unable to stand erect, when my eye caught something
which brought a gush of hope back to my heart.

"I have said that though the floor and ceiling were of iron, the walls were of wood. As I gave a
last hurried glance around, I saw a thin line of yellow light between two of the boards, which
broadened and broadened as a small panel was pushed backward. For an instant I could hardly believe
that here was indeed a door which led away from death. The next instant I threw myself through, and
lay half-fainting upon the other side. The panel had closed again behind me, but the crash of the
lamp, and a few moments afterwards the clang of the two slabs of metal, told me how narrow had been
my escape.