29 Larch Walk
"'Dear me!' he said gravely, 'that is very serious indeed! I am sorry to hear you say that. The fund
was, of course, for the propagation and spread of the red-heads as well as for their maintenance. It
is exceedingly unfortunate that you should be a bachelor.'

"My face lengthened at this, Mr. Holmes, for I thought that I was not to have the vacancy after all;
but after thinking it over for a few minutes he said that it would be all right.

"'In the case of another,' said he, 'the objection might be fatal, but we must stretch a point in
favour of a man with such a head of hair as yours. When shall you be able to enter upon your new
duties?'

"'Well, it is a little awkward, for I have a business already,' said I.

"'Oh, never mind about that, Mr. Wilson!' said Vincent Spaulding. 'I should be able to look after
that for you.'

"'What would be the hours?' I asked.