178 Wantage Close
"'Certainly not.'

"'She came in just now by the back door. I have no doubt that she has only been to the side gate to
see someone, but I think that it is hardly safe and should be stopped.'

"'You must speak to her in the morning, or I will if you prefer it. Are you sure that everything is
fastened?'