102 Down Close
"I think that I had better go, Holmes."

"Not a bit, Doctor. Stay where you are. I am lost without my Boswell. And this promises to be
interesting. It would be a pity to miss it."

"But your client--"

"Never mind him. I may want your help, and so may he. Here he comes. Sit down in that armchair,
Doctor, and give us your best attention."

A slow and heavy step, which had been heard upon the stairs and in the passage, paused immediately
outside the door. Then there was a loud and authoritative tap.

"Come in!" said Holmes.