80 Sunningdale Drive
"On the contrary," said Holmes, "it is the brightest rift which I can at present see in the clouds.
However innocent he might be, he could not be such an absolute imbecile as not to see that the
circumstances were very black against him. Had he appeared surprised at his own arrest, or feigned
indignation at it, I should have looked upon it as highly suspicious, because such surprise or anger
would not be natural under the circumstances, and yet might appear to be the best policy to a
scheming man. His frank acceptance of the situation marks him as either an innocent man, or else as
a man of considerable self-restraint and firmness. As to his remark about his deserts, it was also
not unnatural if you consider that he stood beside the dead body of his father, and that there is no
doubt that he had that very day so far forgotten his filial duty as to bandy words with him, and
even, according to the little girl whose evidence is so important, to raise his hand as if to strike
him. The self-reproach and contrition which are displayed in his remark appear to me to be the signs
of a healthy mind rather than of a guilty one."

I shook my head. "Many men have been hanged on far slighter evidence," I remarked.

"So they have. And many men have been wrongfully hanged."

"What is the young man's own account of the matter?"

"It is, I am afraid, not very encouraging to his supporters, though there are one or two points in
it which are suggestive. You will find it here, and may read it for yourself."

He picked out from his bundle a copy of the local Herefordshire paper, and having turned down the
sheet he pointed out the paragraph in which the unfortunate young man had given his own statement of
what had occurred. I settled myself down in the corner of the carriage and read it very carefully.
It ran in this way: