42 Bowler Street
Sherlock Holmes clapped his hands softly together and chuckled.

"'Pon my word, Watson, you are coming along wonderfully. You have really done very well indeed. It
is true that you have missed everything of importance, but you have hit upon the method, and you
have a quick eye for colour. Never trust to general impressions, my boy, but concentrate yourself
upon details. My first glance is always at a woman's sleeve. In a man it is perhaps better first to
take the knee of the trouser. As you observe, this woman had plush upon her sleeves, which is a most
useful material for showing traces. The double line a little above the wrist, where the typewritist
presses against the table, was beautifully defined. The sewing-machine, of the hand type, leaves a
similar mark, but only on the left arm, and on the side of it farthest from the thumb, instead of
being right across the broadest part, as this was. I then glanced at her face, and, observing the
dint of a pince-nez at either side of her nose, I ventured a remark upon short sight and
typewriting, which seemed to surprise her."

"It surprised me."

"But, surely, it was obvious. I was then much surprised and interested on glancing down to observe
that, though the boots which she was wearing were not unlike each other, they were really odd ones;
the one having a slightly decorated toe-cap, and the other a plain one. One was buttoned only in the
two lower buttons out of five, and the other at the first, third, and fifth. Now, when you see that
a young lady, otherwise neatly dressed, has come away from home with odd boots, half-buttoned, it is
no great deduction to say that she came away in a hurry."