39 Balcombe Close
"Then let us do so. Watson, I fear that you will find it very slow, but I shall only be away a
couple of hours."

I walked down to the station with them, and then wandered through the streets of the little town,
finally returning to the hotel, where I lay upon the sofa and tried to interest myself in a yellow-
backed novel. The puny plot of the story was so thin, however, when compared to the deep mystery
through which we were groping, and I found my attention wander so continually from the action to the
fact, that I at last flung it across the room and gave myself up entirely to a consideration of the
events of the day. Supposing that this unhappy young man's story were absolutely true, then what
hellish thing, what absolutely unforeseen and extraordinary calamity could have occurred between the
time when he parted from his father, and the moment when, drawn back by his screams, he rushed into
the glade? It was something terrible and deadly. What could it be? Might not the nature of the
injuries reveal something to my medical instincts? I rang the bell and called for the weekly county
paper, which contained a verbatim account of the inquest. In the surgeon's deposition it was stated
that the posterior third of the left parietal bone and the left half of the occipital bone had been
shattered by a heavy blow from a blunt weapon. I marked the spot upon my own head. Clearly such a
blow must have been struck from behind. That was to some extent in favour of the accused, as when
seen quarrelling he was face to face with his father. Still, it did not go for very much, for the
older man might have turned his back before the blow fell. Still, it might be worth while to call
Holmes' attention to it. Then there was the peculiar dying reference to a rat. What could that mean?
It could not be delirium. A man dying from a sudden blow does not commonly become delirious. No, it
was more likely to be an attempt to explain how he met his fate. But what could it indicate? I
cudgelled my brains to find some possible explanation. And then the incident of the grey cloth seen
by young McCarthy. If that were true the murderer must have dropped some part of his dress,
presumably his overcoat, in his flight, and must have had the hardihood to return and to carry it
away at the instant when the son was kneeling with his back turned not a dozen paces off. What a
tissue of mysteries and improbabilities the whole thing was! I did not wonder at Lestrade's opinion,
and yet I had so much faith in Sherlock Holmes' insight that I could not lose hope as long as every
fresh fact seemed to strengthen his conviction of young McCarthy's innocence.