40 Tamworth Drive
"Oh! you have seen inside, then?"

"Only once, some years ago. I remember that it was full of papers."

"There isn't a cat in it, for example?"

"No. What a strange idea!"

"Well, look at this!" He took up a small saucer of milk which stood on the top of it.

"No; we don't keep a cat. But there is a cheetah and a baboon."