117 Larch Walk
"Was she in good spirits?"

"Never better. She kept talking of what we should do in our future lives."

"Indeed! That is very interesting. And on the morning of the wedding?"

"She was as bright as possible--at least until after the ceremony."

"And did you observe any change in her then?"

"Well, to tell the truth, I saw then the first signs that I had ever seen that her temper was just a
little sharp. The incident however, was too trivial to relate and can have no possible bearing upon
the case."