42 Wantage Close
"And yet I am not convinced of it," I answered. "The cases which come to light in the papers are, as
a rule, bald enough, and vulgar enough. We have in our police reports realism pushed to its extreme
limits, and yet the result is, it must be confessed, neither fascinating nor artistic."

"A certain selection and discretion must be used in producing a realistic effect," remarked Holmes.
"This is wanting in the police report, where more stress is laid, perhaps, upon the platitudes of
the magistrate than upon the details, which to an observer contain the vital essence of the whole
matter. Depend upon it, there is nothing so unnatural as the commonplace."