88 Larch Walk
"I will do so, and I must be quick, for I have promised Mr. Rucastle to be back before three. I got
his leave to come into town this morning, though he little knew for what purpose."

"Let us have everything in its due order." Holmes thrust his long thin legs out towards the fire and
composed himself to listen.

"In the first place, I may say that I have met, on the whole, with no actual ill-treatment from Mr.
and Mrs. Rucastle. It is only fair to them to say that. But I cannot understand them, and I am not
easy in my mind about them."

"What can you not understand?"

"Their reasons for their conduct. But you shall have it all just as it occurred. When I came down,
Mr. Rucastle met me here and drove me in his dog-cart to the Copper Beeches. It is, as he said,
beautifully situated, but it is not beautiful in itself, for it is a large square block of a house,
whitewashed, but all stained and streaked with damp and bad weather. There are grounds round it,
woods on three sides, and on the fourth a field which slopes down to the Southampton highroad, which
curves past about a hundred yards from the front door. This ground in front belongs to the house,
but the woods all round are part of Lord Southerton's preserves. A clump of copper beeches
immediately in front of the hall door has given its name to the place.