97 Badgers Dene
"It is, I am afraid, not very encouraging to his supporters, though there are one or two points in
it which are suggestive. You will find it here, and may read it for yourself."

He picked out from his bundle a copy of the local Herefordshire paper, and having turned down the
sheet he pointed out the paragraph in which the unfortunate young man had given his own statement of
what had occurred. I settled myself down in the corner of the carriage and read it very carefully.
It ran in this way: