120 Beaconside
"Subtle enough and horrible enough. When a doctor does go wrong he is the first of criminals. He has
nerve and he has knowledge. Palmer and Pritchard were among the heads of their profession. This man
strikes even deeper, but I think, Watson, that we shall be able to strike deeper still. But we shall
have horrors enough before the night is over; for goodness' sake let us have a quiet pipe and turn
our minds for a few hours to something more cheerful."

About nine o'clock the light among the trees was extinguished, and all was dark in the direction of
the Manor House. Two hours passed slowly away, and then, suddenly, just at the stroke of eleven, a
single bright light shone out right in front of us.

"That is our signal," said Holmes, springing to his feet; "it comes from the middle window."

As we passed out he exchanged a few words with the landlord, explaining that we were going on a late
visit to an acquaintance, and that it was possible that we might spend the night there. A moment
later we were out on the dark road, a chill wind blowing in our faces, and one yellow light
twinkling in front of us through the gloom to guide us on our sombre errand.