52 Wilson Gardens
"And that is--"

"That McCarthy senior met his death from McCarthy junior and that all theories to the contrary are
the merest moonshine."

"Well, moonshine is a brighter thing than fog," said Holmes, laughing. "But I am very much mistaken
if this is not Hatherley Farm upon the left."

"Yes, that is it." It was a widespread, comfortable-looking building, two-storied, slate-roofed,
with great yellow blotches of lichen upon the grey walls. The drawn blinds and the smokeless
chimneys, however, gave it a stricken look, as though the weight of this horror still lay heavy upon
it. We called at the door, when the maid, at Holmes' request, showed us the boots which her master
wore at the time of his death, and also a pair of the son's, though not the pair which he had then
had. Having measured these very carefully from seven or eight different points, Holmes desired to be
led to the court-yard, from which we all followed the winding track which led to Boscombe Pool.

Sherlock Holmes was transformed when he was hot upon such a scent as this. Men who had only known
the quiet thinker and logician of Baker Street would have failed to recognise him. His face flushed
and darkened. His brows were drawn into two hard black lines, while his eyes shone out from beneath
them with a steely glitter. His face was bent downward, his shoulders bowed, his lips compressed,
and the veins stood out like whipcord in his long, sinewy neck. His nostrils seemed to dilate with a
purely animal lust for the chase, and his mind was so absolutely concentrated upon the matter before
him that a question or remark fell unheeded upon his ears, or, at the most, only provoked a quick,
impatient snarl in reply. Swiftly and silently he made his way along the track which ran through the
meadows, and so by way of the woods to the Boscombe Pool. It was damp, marshy ground, as is all that
district, and there were marks of many feet, both upon the path and amid the short grass which
bounded it on either side. Sometimes Holmes would hurry on, sometimes stop dead, and once he made
quite a little detour into the meadow. Lestrade and I walked behind him, the detective indifferent
and contemptuous, while I watched my friend with the interest which sprang from the conviction that
every one of his actions was directed towards a definite end.