2 Moor End
"Now for Mr. Breckinridge," he continued, buttoning up his coat as we came out into the frosty air.
"Remember, Watson that though we have so homely a thing as a goose at one end of this chain, we have
at the other a man who will certainly get seven years' penal servitude unless we can establish his
innocence. It is possible that our inquiry may but confirm his guilt; but, in any case, we have a
line of investigation which has been missed by the police, and which a singular chance has placed in
our hands. Let us follow it out to the bitter end. Faces to the south, then, and quick march!"

We passed across Holborn, down Endell Street, and so through a zigzag of slums to Covent Garden
Market. One of the largest stalls bore the name of Breckinridge upon it, and the proprietor a
horsey-looking man, with a sharp face and trim side-whiskers was helping a boy to put up the
shutters.

"Good-evening. It's a cold night," said Holmes.

The salesman nodded and shot a questioning glance at my companion.