12 Larch Walk
A loud thudding noise came from somewhere downstairs. "That is Mrs. Toller in the cellar," said she.
"Her husband lies snoring on the kitchen rug. Here are his keys, which are the duplicates of Mr.
Rucastle's."

"You have done well indeed!" cried Holmes with enthusiasm. "Now lead the way, and we shall soon see
the end of this black business."