116 Badgers Dene
"He was a very shy man, Mr. Holmes. He would rather walk with me in the evening than in the
daylight, for he said that he hated to be conspicuous. Very retiring and gentlemanly he was. Even
his voice was gentle. He'd had the quinsy and swollen glands when he was young, he told me, and it
had left him with a weak throat, and a hesitating, whispering fashion of speech. He was always well
dressed, very neat and plain, but his eyes were weak, just as mine are, and he wore tinted glasses
against the glare."

"Well, and what happened when Mr. Windibank, your stepfather, returned to France?"

"Mr. Hosmer Angel came to the house again and proposed that we should marry before father came back.
He was in dreadful earnest and made me swear, with my hands on the Testament, that whatever happened
I would always be true to him. Mother said he was quite right to make me swear, and that it was a
sign of his passion. Mother was all in his favour from the first and was even fonder of him than I
was. Then, when they talked of marrying within the week, I began to ask about father; but they both
said never to mind about father, but just to tell him afterwards, and mother said she would make it
all right with him. I didn't quite like that, Mr. Holmes. It seemed funny that I should ask his
leave, as he was only a few years older than me; but I didn't want to do anything on the sly, so I
wrote to father at Bordeaux, where the company has its French offices, but the letter came back to
me on the very morning of the wedding."

"It missed him, then?"

"Yes, sir; for he had started to England just before it arrived."

"Ha! that was unfortunate. Your wedding was arranged, then, for the Friday. Was it to be in church?"