112 Bowler Street
"An elderly man, I presume?" said Holmes.

"About sixty; but his constitution has been shattered by his life abroad, and he has been in failing
health for some time. This business has had a very bad effect upon him. He was an old friend of
McCarthy's, and, I may add, a great benefactor to him, for I have learned that he gave him Hatherley
Farm rent free."

"Indeed! That is interesting," said Holmes.

"Oh, yes! In a hundred other ways he has helped him. Everybody about here speaks of his kindness to
him."

"Really! Does it not strike you as a little singular that this McCarthy, who appears to have had
little of his own, and to have been under such obligations to Turner, should still talk of marrying
his son to Turner's daughter, who is, presumably, heiress to the estate, and that in such a very
cocksure manner, as if it were merely a case of a proposal and all else would follow? It is the more
strange, since we know that Turner himself was averse to the idea. The daughter told us as much. Do
you not deduce something from that?"