74 Harvey Street
"'The family of Lord Robert St. Simon has been thrown into the greatest consternation by the strange
and painful episodes which have taken place in connection with his wedding. The ceremony, as shortly
announced in the papers of yesterday, occurred on the previous morning; but it is only now that it
has been possible to confirm the strange rumours which have been so persistently floating about. In
spite of the attempts of the friends to hush the matter up, so much public attention has now been
drawn to it that no good purpose can be served by affecting to disregard what is a common subject
for conversation. "'The ceremony, which was performed at St. George's, Hanover Square, was a very
quiet one, no one being present save the father of the bride, Mr. Aloysius Doran, the Duchess of
Balmoral, Lord Backwater, Lord Eustace and Lady Clara St. Simon (the younger brother and sister of
the bridegroom), and Lady Alicia Whittington. The whole party proceeded afterwards to the house of
Mr. Aloysius Doran, at Lancaster Gate, where breakfast had been prepared. It appears that some
little trouble was caused by a woman, whose name has not been ascertained, who endeavoured to force
her way into the house after the bridal party, alleging that she had some claim upon Lord St. Simon.
It was only after a painful and prolonged scene that she was ejected by the butler and the footman.
The bride, who had fortunately entered the house before this unpleasant interruption, had sat down
to breakfast with the rest, when she complained of a sudden indisposition and retired to her room.
Her prolonged absence having caused some comment, her father followed her, but learned from her maid
that she had only come up to her chamber for an instant, caught up an ulster and bonnet, and hurried
down to the passage. One of the footmen declared that he had seen a lady leave the house thus
apparelled, but had refused to credit that it was his mistress, believing her to be with the
company. On ascertaining that his daughter had disappeared, Mr. Aloysius Doran, in conjunction with
the bridegroom, instantly put themselves in communication with the police, and very energetic
inquiries are being made, which will probably result in a speedy clearing up of this very singular
business. Up to a late hour last night, however, nothing had transpired as to the whereabouts of the
missing lady. There are rumours of foul play in the matter, and it is said that the police have
caused the arrest of the woman who had caused the original disturbance, in the belief that, from
jealousy or some other motive, she may have been concerned in the strange disappearance of the
bride.'"

"And is that all?"