19 Tamworth Drive
"I think that there you ask a little too much," responded his Lordship. "I may be forced to
acquiesce in these recent developments, but I can hardly be expected to make merry over them. I
think that with your permission I will now wish you all a very good-night." He included us all in a
sweeping bow and stalked out of the room.

"Then I trust that you at least will honour me with your company," said Sherlock Holmes. "It is
always a joy to meet an American, Mr. Moulton, for I am one of those who believe that the folly of a
monarch and the blundering of a minister in far-gone years will not prevent our children from being
some day citizens of the same world-wide country under a flag which shall be a quartering of the
Union Jack with the Stars and Stripes."