5 Wilson Gardens
The lady coloured deeply and covered over her injured wrist. "He is a hard man," she said, "and
perhaps he hardly knows his own strength."

There was a long silence, during which Holmes leaned his chin upon his hands and stared into the
crackling fire.

"This is a very deep business," he said at last. "There are a thousand details which I should desire
to know before I decide upon our course of action. Yet we have not a moment to lose. If we were to
come to Stoke Moran to-day, would it be possible for us to see over these rooms without the
knowledge of your stepfather?"