73 Moor End
"One day a gold convoy came down from Ballarat to Melbourne, and we lay in wait for it and attacked
it. There were six troopers and six of us, so it was a close thing, but we emptied four of their
saddles at the first volley. Three of our boys were killed, however, before we got the swag. I put
my pistol to the head of the wagon-driver, who was this very man McCarthy. I wish to the Lord that I
had shot him then, but I spared him, though I saw his wicked little eyes fixed on my face, as though
to remember every feature. We got away with the gold, became wealthy men, and made our way over to
England without being suspected. There I parted from my old pals and determined to settle down to a
quiet and respectable life. I bought this estate, which chanced to be in the market, and I set
myself to do a little good with my money, to make up for the way in which I had earned it. I
married, too, and though my wife died young she left me my dear little Alice. Even when she was just
a baby her wee hand seemed to lead me down the right path as nothing else had ever done. In a word,
I turned over a new leaf and did my best to make up for the past. All was going well when McCarthy
laid his grip upon me.

"I had gone up to town about an investment, and I met him in Regent Street with hardly a coat to his
back or a boot to his foot.