12 Harvey Street
"I have no doubt that I am very stupid, but I must confess that I am unable to follow you. For
example, how did you deduce that this man was intellectual?"

For answer Holmes clapped the hat upon his head. It came right over the forehead and settled upon
the bridge of his nose. "It is a question of cubic capacity," said he; "a man with so large a brain
must have something in it."