98 Beaconside
"I took it up and examined it. It was of the same peculiar tint, and the same thickness. But then
the impossibility of the thing obtruded itself upon me. How could my hair have been locked in the
drawer? With trembling hands I undid my trunk, turned out the contents, and drew from the bottom my
own hair. I laid the two tresses together, and I assure you that they were identical. Was it not
extraordinary? Puzzle as I would, I could make nothing at all of what it meant. I returned the
strange hair to the drawer, and I said nothing of the matter to the Rucastles as I felt that I had
put myself in the wrong by opening a drawer which they had locked.

"I am naturally observant, as you may have remarked, Mr. Holmes, and I soon had a pretty good plan
of the whole house in my head. There was one wing, however, which appeared not to be inhabited at
all. A door which faced that which led into the quarters of the Tollers opened into this suite, but
it was invariably locked. One day, however, as I ascended the stair, I met Mr. Rucastle coming out
through this door, his keys in his hand, and a look on his face which made him a very different
person to the round, jovial man to whom I was accustomed. His cheeks were red, his brow was all
crinkled with anger, and the veins stood out at his temples with passion. He locked the door and
hurried past me without a word or a look.

"This aroused my curiosity, so when I went out for a walk in the grounds with my charge, I strolled
round to the side from which I could see the windows of this part of the house. There were four of
them in a row, three of which were simply dirty, while the fourth was shuttered up. They were
evidently all deserted. As I strolled up and down, glancing at them occasionally, Mr. Rucastle came
out to me, looking as merry and jovial as ever.

"'Ah!' said he, 'you must not think me rude if I passed you without a word, my dear young lady. I
was preoccupied with business matters.'