64 Harvey Street
"To an end?"

"Yes, sir. And no later than this morning. I went to my work as usual at ten o'clock, but the door
was shut and locked, with a little square of cardboard hammered on to the middle of the panel with a
tack. Here it is, and you can read for yourself."