24 Beaconside
"I should be very glad now to go upstairs," said he. "I shall probably wish to go over the outside
of the house again. Perhaps I had better take a look at the lower windows before I go up."

He walked swiftly round from one to the other, pausing only at the large one which looked from the
hall onto the stable lane. This he opened and made a very careful examination of the sill with his
powerful magnifying lens. "Now we shall go upstairs," said he at last.

The banker's dressing-room was a plainly furnished little chamber, with a grey carpet, a large
bureau, and a long mirror. Holmes went to the bureau first and looked hard at the lock.

"Which key was used to open it?" he asked.

"That which my son himself indicated--that of the cupboard of the lumber-room."

"Have you it here?"