96 Wantage Close
"'Oh, yes,' said he, turning to me, 'we are very much obliged to you, Miss Hunter, for falling in
with our whims so far as to cut your hair. I assure you that it has not detracted in the tiniest
iota from your appearance. We shall now see how the electric-blue dress will become you. You will
find it laid out upon the bed in your room, and if you would be so good as to put it on we should
both be extremely obliged.'

"The dress which I found waiting for me was of a peculiar shade of blue. It was of excellent
material, a sort of beige, but it bore unmistakable signs of having been worn before. It could not
have been a better fit if I had been measured for it. Both Mr. and Mrs. Rucastle expressed a delight
at the look of it, which seemed quite exaggerated in its vehemence. They were waiting for me in the
drawing-room, which is a very large room, stretching along the entire front of the house, with three
long windows reaching down to the floor. A chair had been placed close to the central window, with
its back turned towards it. In this I was asked to sit, and then Mr. Rucastle, walking up and down
on the other side of the room, began to tell me a series of the funniest stories that I have ever
listened to. You cannot imagine how comical he was, and I laughed until I was quite weary. Mrs.
Rucastle, however, who has evidently no sense of humour, never so much as smiled, but sat with her
hands in her lap, and a sad, anxious look upon her face. After an hour or so, Mr. Rucastle suddenly
remarked that it was time to commence the duties of the day, and that I might change my dress and go
to little Edward in the nursery.

"Two days later this same performance was gone through under exactly similar circumstances. Again I
changed my dress, again I sat in the window, and again I laughed very heartily at the funny stories
of which my employer had an immense répertoire, and which he told inimitably. Then he handed me a
yellow-backed novel, and moving my chair a little sideways, that my own shadow might not fall upon
the page, he begged me to read aloud to him. I read for about ten minutes, beginning in the heart of
a chapter, and then suddenly, in the middle of a sentence, he ordered me to cease and to change my
dress.