37 Harvey Street
"I have no doubt that I am very stupid, but I must confess that I am unable to follow you. For
example, how did you deduce that this man was intellectual?"

For answer Holmes clapped the hat upon his head. It came right over the forehead and settled upon
the bridge of his nose. "It is a question of cubic capacity," said he; "a man with so large a brain
must have something in it."

"The decline of his fortunes, then?"

"This hat is three years old. These flat brims curled at the edge came in then. It is a hat of the
very best quality. Look at the band of ribbed silk and the excellent lining. If this man could
afford to buy so expensive a hat three years ago, and has had no hat since, then he has assuredly
gone down in the world."